Authentication and Authorization In JWT

+++ controllers/api/v1/posts_api.js
@@ -21,16 +21,17 @@ module.exports.index = async function (req, res) {
 module.exports.destroy = async function (req, res) {
   try {
     let post = await Post.findById(req.params.id);
-    // if (post.user == req.user.id) {
-    post.remove();
-    await Comment.deleteMany({ post: req.params.id });
-    return res.json(200, {
-      message: "Post and assosciated comments deleted successfully",
-    });
-    // } else {
-    //   req.flash("success", "You can not delete this post !");
-    //   return res.redirect("back");
-    // }
+    if (post.user == req.user.id) {
+      post.remove();
+      await Comment.deleteMany({ post: req.params.id });
+      return res.json(200, {
+        message: "Post and assosciated comments deleted successfully",
+      });
+    } else {
+      return res.json(401, {
+        message: "you can not delete this post",
+      });
+    }
   } catch (err) {
     // console.log("error in creating in post");
     console.log(err);
